refactor(nx-governance): separate snapshot drift contracts
