+++ packages/governance/src/core/index.ts
@@ -2,9 +2,11 @@ export * from './adapter.js';
 export * from './assessment.js';
 export * from './built-in-rule-pack.js';
 export * from './built-in-rules.js';
+export * from './drift.js';
 export * from './models.js';
 export * from './exceptions.js';
 export * from './profile.js';
 export * from './rule-engine.js';
 export * from './rules.js';
 export * from './signals.js';
+export * from './snapshots.js';
